ensure az and gh cli before running gh setup
